requirements.txt
streamlit==1.30.0
anthropic==0.18.1
langgraph==0.0.32
langchain==0.1.9
langchain-anthropic==0.1.4
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
folium==0.15.1
plotly==5.18.0
streamlit-folium==0.15.1
pandas==2.1.4
openmeteo-requests==1.1.0
requests-cache==1.1.1
retry-requests==2.0.0